Route group message delete &  newslink delete back to correct group news page

+++ client/app/components/News/SocialLink/SocialLinkListItem/index.js
@@ -64,7 +64,8 @@ export function SocialLinkListItem(props) {
         <Button
           size='small'
           onClick={() => {
-            props.deleteSocialLinkBegin(newsItem.social_link);
+            if (confirm('Delete social link?'))
+              props.deleteSocialLinkBegin(newsItem.social_link);
           }}
         >
           Delete
